requirements.txt
pyairtable
python-dotenv
requests
pydantic
pytest
loguru
google.generativeai
tenacity
pyYaml